Fixed locations of the classes
